+++ R/googleAuthR_discovery.R
@@ -254,7 +254,7 @@ gar_create_api_objects <- function(filename = "./inst/api_objects.R", api_json){
   
   lapply(paste(fd, fp,fb, sep = "\n\n"), add_line, temp)
   
-  suppressMessages(formatR::tidy_eval(temp, file = filename, width.cutoff = 80))
+  formatR::tidy_eval(temp, file = filename, width.cutoff = 80)
   
 }