+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Add client based authentication in JavaScript plus example app
 * Add check to `gar_auth_service` to see if you have downloaded right JSON file
+* Change warnings() in batch to myMessage() level 2
 
 # v0.3.1 - CRAN