+++ NEWS.md
@@ -6,6 +6,7 @@
 * Add check to `gar_auth_service` to see if you have downloaded right JSON file
 * Discovery API functions to get details on Google APIs added: `gar_discovery_apis_list` and `gar_discovery_api`
 * Change warnings() in batch to myMessage() level 2
+* ensure batch requests only occur per second to help calculation of QPS limits
 
 # v0.3.1 - CRAN