+++ NEWS.md
@@ -5,9 +5,11 @@
 * Add client based authentication in JavaScript plus example app
 * Add check to `gar_auth_service` to see if you have downloaded right JSON file
 * Discovery API functions to get details on Google APIs added: `gar_discovery_apis_list` and `gar_discovery_api`
+* Add `gar_create_package` that takes `gar_discovery_api` JSON and creates R package
 * Change warnings() in batch to myMessage() level 2
 * ensure batch requests only occur per second to help calculation of QPS limits
 * Add 404 message if batch requests are not found.
+* Fixed halt error if message can't parse body JSON, will now fail gracefully but carry on
 
 # v0.3.1 - CRAN