+++ NEWS.md
@@ -13,6 +13,7 @@
 * allow overwriting of default httr "encode" again (#28)
 * Headers will contain up to date version number of package
 * Add `gar_auto_auth` and `gar_attach_auto_auth` for auto-authentication upon a package load
+* Fix bug where you couldn't pass in the file location of the ".httr-oauth" location to `gar_auth()`
 
 # v0.3.1 - CRAN